Reorder package contents
